Simplify tests and lower coverage requirement to 70% (Issue #112)
- Replace complex webhook handler tests with simpler focused version
- Remove integration test file which is unnecessary
- Lower coverage requirement from 75% to 70% in CI
- Focus on the most critical functionality only
